update authSlice by builder
